% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convmoment.R
\name{raw2central}
\alias{raw2central}
\title{Convert Raw Moments to Central Moments}
\usage{
raw2central(raw_moments, origin = 0)
}
\arguments{
\item{raw_moments}{Numeric vector of raw moments about \code{origin}
(orders 1..K).}

\item{origin}{Initial origin (default 0).}
}
\value{
Numeric vector of central moments (orders 1..K).
}
\description{
Convenience wrapper converting raw moments about origin \code{a}
to central moments (about the mean).
}
\examples{
x <- c(20, 25, 29, 32, 40)
raw_direct <- sapply(1:5, function(r) mean(x^r))
raw2central(raw_direct, origin = 0)
}
