% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convmoment.R
\name{conv_moment_all}
\alias{conv_moment_all}
\title{Moment Conversion}
\usage{
conv_moment_all(x, a, k)
}
\arguments{
\item{x}{A numeric vector with moments about \code{a}, the existing origin.}

\item{a}{The old origin}

\item{k}{The new origin. To convert to central moments, use \eqn{\mu_1'+a}}
}
\value{
A vector of converted moments
}
\description{
This function converts moments of a distribution to and from central and
raw ones. Raw moments about a origin can be converted to raw moments with
another origin.
}
\details{
If you want to convert to central moment, use the same function. Just
remember to use the arithmetic mean as the new origin.

Mean = First raw moment about \code{a} + \code{a}; where \code{a} is the old origin.
}
\examples{
x <- c(1, 16, -40)
conv_moment_all(x, 2, 0)
}
