% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convmoment-package.R
\docType{package}
\name{convmoment-package}
\alias{convmoment}
\alias{convmoment-package}
\title{convmoment: Converts moments of a distribution to and from central and raw ones}
\description{
The package converts to and from raw moments to and from other raw moments or central moments.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://www.thinkermahmud.com/convmoment/}
}

}
\author{
\strong{Maintainer}: Abdullah Al Mahmud \email{almahmud.sbi@gmail.com} (\href{https://orcid.org/0000-0003-2814-8798}{ORCID})

}
\keyword{internal}
