% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_moments.R
\name{get_moments}
\alias{get_moments}
\title{Moment Estimation}
\usage{
get_moments(x, a, r, decimal = 3)
}
\arguments{
\item{x}{A numeric vector with values of a variable}

\item{a}{The origin}

\item{r}{The highest order}

\item{decimal}{The no. of digits after decimal}
}
\value{
A data frame of \code{r} rows and 2 columns; the columns represent
order of moment and the moment.
}
\description{
This function estimates raw moments around a certain origin or central
moments up to a certain degree (r)
}
\details{
To find the raw moments use the arbitrary origin \code{a}, while, for
central moments, use \code{a = mean(x)}.
}
\examples{
x <- sample(10, 5)
get_moments(x, 2, 3) # test r > 0
}
